agent-blueprint/fargate-mcp-farm/nova-act-mcp/src/requirements.txt
fastmcp>=2.11.0
nova-act
python-dotenv
psutil
fastapi
uvicorn[standard]
httpx
playwright